requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
coverage==7.3.2
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
flake8==6.1.0
mccabe==0.7.0
Pillow==10.1.0
psycopg2==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3
